+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 :warning: **Important**: If you are upgrading to version 2.3.0 or later and have firewall rules or network configuration that blocks any unknown traffic by default, you need to update your configuration to allow connections to the new DNS names and IP addresses. Please refer to this [changelog](#230-january-23-2023) for more details.
 
+2.12.5 (WIP)
+============
+
+Bug Fixes
+---------
+- Fixed a Citrix issue where calling `device.audio.setInputDevice()`, results in a `MediaStreamError`.
+  - User must add `createMediaStream` to their `Device.Options`.
+  - [Device.Options.createMediaStream](https://twilio.github.io/twilio-voice.js/interfaces/voice.device.options.html#createMediaStream)
+
 2.12.4 (March 12, 2025)
 =======================